requirements.txt
numpy==1.19.5
opencv-python==4.5.3.56
opencv-contrib-python==4.5.3.56
scikit-learn==1.0.2
scikit-image==0.18.3
scipy==1.7.3
setuptools==63.4.1
importlib-metadata==4.12.0
ipykernel==6.9.1
ipython==7.31.1
ipywidgets==7.6.3
jedi==0.18.1
Jinja2==3.0.1
jsonschema==4.9.1
jupyter==1.0.0
jupyter-client==7.2.2
jupyter-console==6.4.0
jupyter-core==4.10.0
matplotlib==3.5.2
more-itertools==8.8.0
notebook==6.4.0
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
prompt-toolkit==3.0.30
protobuf==3.20.1
ptyprocess==0.7.0
pydot==1.4.2
Pygments==2.12.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
python-osc==1.8.0
pytz==2022.1
PyYAML==5.4.1
qtconsole==5.2.2
termcolor==1.1.0
terminado==0.13.3
testpath==0.6.0
tornado==6.2
tqdm==4.64.0
wcwidth==0.2.5
wheel==0.37.1
wrapt==1.12.1
tensorflow==2.5.1
seaborn==0.11.2
labelImg
pyserial==3.5
lpips==0.1.4
PyMuPDF==1.20.1
requests==2.28.1
flirpy==0.3.0
syphonpy==0.0.2
glfw==2.5.4
PyOpenGL==3.1.5
PyOpenGL_accelerate==3.1.5
sh==1.14.2
gphoto2==2.3.4 --only-binary :all:
